World View Projection 적용, Camera 컴포넌트 추가

+++ Engine/Entity.cpp
@@ -4,13 +4,13 @@ namespace arias
 {
 	Entity::Entity() :
 		mName{},
-		mID(reinterpret_cast<UINT32>(this))
+		mID(reinterpret_cast<UINT64>(this))
 	{
 	}
 
 	Entity::Entity(const Entity& other) :
 		mName(other.mName),
-		mID(reinterpret_cast<UINT32>(this))
+		mID(reinterpret_cast<UINT64>(this))
 	{
 	}
